backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio[asgi]==5.12.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.14.1
redis==5.2.1
motor==3.6.1
celery==5.4.0
opencv-python-headless==4.10.0.84
numpy==2.2.2
scikit-learn==1.6.1
torch==2.5.1
ultralytics==8.3.68
httpx==0.28.1
tenacity==9.0.0
structlog==25.1.0
prometheus-client==0.21.1
pytest==8.3.4
pytest-asyncio==0.25.3
